rename navbar to SluggoNavbar/. move NavbarTeamPageLink to own directory.
