+++ src/components/NavbarTeamPageLink/index.ts
@@ -1,0 +1,3 @@
+import NavbarTeamPageLink from "./NavbarTeamPageLink.vue";
+
+export default NavbarTeamPageLink;
