+++ src/components/SluggoNavbar/index.ts
@@ -1,0 +1,3 @@
+import SluggoNavbar from "./SluggoNavbar.vue";
+
+export default SluggoNavbar;
