+++ src/components/SluggoNavbar/styles.module.scss
@@ -1,0 +1,4 @@
+.navigationContainer {
+  display: grid;
+  grid-template-columns: 1fr 2fr;
+}
