--- src/components/navbar/index.ts
@@ -1,3 +1,0 @@
-import SluggoNavbar from "@/components/navbar/SluggoNavbar.vue";
-
-export default SluggoNavbar;
