+++ src/router/index.ts
@@ -7,13 +7,13 @@ import Help from "@/views/Help.vue";
 import Auth from "@/views/Auth.vue";
 import Profile from "@/views/Profile.vue";
 import NewTeam from "@/views/NewTeam.vue";
-import TeamWrapper from "@/components/TeamWrapper.vue";
+import SluggoNavbar from "@/components/SluggoNavbar";
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: "/teams/:teamId",
     name: "Team",
-    component: TeamWrapper,
+    component: SluggoNavbar,
     children: [
       {
         path: "",
